% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_confirm_sign_up}
\alias{admin_confirm_sign_up}
\title{Confirms user registration as an admin without using a confirmation code}
\usage{
admin_confirm_sign_up(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for which you want to confirm user registration.}

\item{Username}{[required] The user name for which you want to confirm user registration.}
}
\description{
Confirms user registration as an admin without using a confirmation code. Works on any user.
}
\details{
Requires developer credentials.
}
\section{Accepted Parameters}{
\preformatted{admin_confirm_sign_up(
  UserPoolId = "string",
  Username = "string"
)
}
}

