% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_disable_user}
\alias{admin_disable_user}
\title{Disables the specified user as an administrator}
\usage{
admin_disable_user(
  UserPoolId = "string",
  Username = "string"
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to disable the user.}

\item{Username}{[required] The user name of the user you wish to disable.}
}
\description{
Disables the specified user as an administrator. Works on any user.
}
\details{
Requires developer credentials.
}
\examples{

}
