% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_enable_user}
\alias{admin_enable_user}
\title{Enables the specified user as an administrator}
\usage{
admin_enable_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool where you want to enable the user.}

\item{Username}{The user name of the user you wish to enable.}
}
\description{
Enables the specified user as an administrator. Works on any user.
}
\details{
Requires developer credentials.
}
\examples{

}
