% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_get_device}
\alias{admin_get_device}
\title{Gets the device, as an administrator}
\usage{
admin_get_device(DeviceKey, UserPoolId, Username)
}
\arguments{
\item{DeviceKey}{The device key.}

\item{UserPoolId}{The user pool ID.}

\item{Username}{The user name.}
}
\description{
Gets the device, as an administrator.
}
\details{
Requires developer credentials.
}
\examples{

}
