% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_get_user}
\alias{admin_get_user}
\title{Gets the specified user by user name in a user pool as an administrator}
\usage{
admin_get_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to get information about the user.}

\item{Username}{[required] The user name of the user you wish to retrieve.}
}
\description{
Gets the specified user by user name in a user pool as an administrator. Works on any user.
}
\details{
Requires developer credentials.
}
\section{Accepted Parameters}{
\preformatted{admin_get_user(
  UserPoolId = "string",
  Username = "string"
)
}
}

