% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_set_user_mfa_preference}
\alias{admin_set_user_mfa_preference}
\title{Sets the user's multi-factor authentication (MFA) preference}
\usage{
admin_set_user_mfa_preference(Username, UserPoolId, SMSMfaSettings = NULL,
  SoftwareTokenMfaSettings = NULL)
}
\arguments{
\item{Username}{The user pool username or alias.}

\item{UserPoolId}{The user pool ID.}

\item{SMSMfaSettings}{The SMS text message MFA settings.}

\item{SoftwareTokenMfaSettings}{The time-based one-time password software token MFA settings.}
}
\description{
Sets the user's multi-factor authentication (MFA) preference.
}
\examples{

}
