% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_set_user_settings}
\alias{admin_set_user_settings}
\title{Sets all the user settings for a specified user name}
\usage{
admin_set_user_settings(UserPoolId, Username, MFAOptions)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool where you want to set the user's settings, such as MFA options.}

\item{Username}{The user name of the user for whom you wish to set user settings.}

\item{MFAOptions}{Specifies the options for MFA (e.g., email or phone number).}
}
\description{
Sets all the user settings for a specified user name. Works on any user.
}
\details{
Requires developer credentials.
}
\examples{

}
