% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_update_device_status}
\alias{admin_update_device_status}
\title{Updates the device status as an administrator}
\usage{
admin_update_device_status(UserPoolId, Username, DeviceKey,
  DeviceRememberedStatus = NULL)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{Username}{[required] The user name.}

\item{DeviceKey}{[required] The device key.}

\item{DeviceRememberedStatus}{The status indicating whether a device has been remembered or not.}
}
\description{
Updates the device status as an administrator.
}
\details{
Requires developer credentials.
}
\section{Accepted Parameters}{
\preformatted{admin_update_device_status(
  UserPoolId = "string",
  Username = "string",
  DeviceKey = "string",
  DeviceRememberedStatus = "remembered"|"not_remembered"
)
}
}

