% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{confirm_forgot_password}
\alias{confirm_forgot_password}
\title{Allows a user to enter a confirmation code to reset a forgotten password}
\usage{
confirm_forgot_password(ClientId, Username, ConfirmationCode, Password,
  SecretHash = NULL, AnalyticsMetadata = NULL, UserContextData = NULL)
}
\arguments{
\item{ClientId}{The app client ID of the app associated with the user pool.}

\item{Username}{The user name of the user for whom you want to enter a code to retrieve a forgotten password.}

\item{ConfirmationCode}{The confirmation code sent by a user's request to retrieve a forgotten password. For more information, see}

\item{Password}{The password sent by a user's request to retrieve a forgotten password.}

\item{SecretHash}{A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.}

\item{AnalyticsMetadata}{The Amazon Pinpoint analytics metadata for collecting metrics for \code{ConfirmForgotPassword} calls.}

\item{UserContextData}{Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.}
}
\description{
Allows a user to enter a confirmation code to reset a forgotten password.
}
\examples{

}
