% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{confirm_sign_up}
\alias{confirm_sign_up}
\title{Confirms registration of a user and handles the existing alias from a previous user}
\usage{
confirm_sign_up(ClientId, Username, ConfirmationCode, SecretHash = NULL,
  ForceAliasCreation = NULL, AnalyticsMetadata = NULL,
  UserContextData = NULL)
}
\arguments{
\item{ClientId}{The ID of the app client associated with the user pool.}

\item{Username}{The user name of the user whose registration you wish to confirm.}

\item{ConfirmationCode}{The confirmation code sent by a user's request to confirm registration.}

\item{SecretHash}{A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.}

\item{ForceAliasCreation}{Boolean to be specified to force user confirmation irrespective of existing alias. By default set to \code{False}. If this parameter is set to \code{True} and the phone number/email used for sign up confirmation already exists as an alias with a different user, the API call will migrate the alias from the previous user to the newly created user being confirmed. If set to \code{False}, the API will throw an \strong{AliasExistsException} error.}

\item{AnalyticsMetadata}{The Amazon Pinpoint analytics metadata for collecting metrics for \code{ConfirmSignUp} calls.}

\item{UserContextData}{Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.}
}
\description{
Confirms registration of a user and handles the existing alias from a previous user.
}
\examples{

}
