% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{create_resource_server}
\alias{create_resource_server}
\title{Creates a new OAuth2}
\usage{
create_resource_server(UserPoolId, Identifier, Name, Scopes = NULL)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool.}

\item{Identifier}{A unique resource server identifier for the resource server. This could be an HTTPS endpoint where the resource server is located. For example, \code{https://my-weather-api.example.com}.}

\item{Name}{A friendly name for the resource server.}

\item{Scopes}{A list of scopes. Each scope is map, where the keys are \code{name} and \code{description}.}
}
\description{
Creates a new OAuth2.0 resource server and defines custom scopes in it.
}
\examples{

}
