% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{create_user_import_job}
\alias{create_user_import_job}
\title{Creates the user import job}
\usage{
create_user_import_job(JobName, UserPoolId, CloudWatchLogsRoleArn)
}
\arguments{
\item{JobName}{[required] The job name for the user import job.}

\item{UserPoolId}{[required] The user pool ID for the user pool that the users are being imported into.}

\item{CloudWatchLogsRoleArn}{[required] The role ARN for the Amazon CloudWatch Logging role for the user import job.}
}
\description{
Creates the user import job.
}
\section{Accepted Parameters}{
\preformatted{create_user_import_job(
  JobName = "string",
  UserPoolId = "string",
  CloudWatchLogsRoleArn = "string"
)
}
}

