% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{delete_group}
\alias{delete_group}
\title{Deletes a group}
\usage{
delete_group(GroupName, UserPoolId)
}
\arguments{
\item{GroupName}{The name of the group.}

\item{UserPoolId}{The user pool ID for the user pool.}
}
\description{
Deletes a group. Currently only groups with no members can be deleted.
}
\details{
Requires developer credentials.
}
\examples{

}
