% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{delete_identity_provider}
\alias{delete_identity_provider}
\title{Deletes an identity provider for a user pool}
\usage{
delete_identity_provider(UserPoolId, ProviderName)
}
\arguments{
\item{UserPoolId}{The user pool ID.}

\item{ProviderName}{The identity provider name.}
}
\description{
Deletes an identity provider for a user pool.
}
\examples{

}
