% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{delete_resource_server}
\alias{delete_resource_server}
\title{Deletes a resource server}
\usage{
delete_resource_server(UserPoolId, Identifier)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool that hosts the resource server.}

\item{Identifier}{The identifier for the resource server.}
}
\description{
Deletes a resource server.
}
\examples{

}
