% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{delete_user}
\alias{delete_user}
\title{Allows a user to delete himself or herself}
\usage{
delete_user(AccessToken)
}
\arguments{
\item{AccessToken}{The access token from a request to delete a user.}
}
\description{
Allows a user to delete himself or herself.
}
\examples{

}
