% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{delete_user_attributes}
\alias{delete_user_attributes}
\title{Deletes the attributes for a user}
\usage{
delete_user_attributes(UserAttributeNames, AccessToken)
}
\arguments{
\item{UserAttributeNames}{An array of strings representing the user attribute names you wish to delete.

For custom attributes, you must prepend the \code{custom:} prefix to the attribute name.}

\item{AccessToken}{The access token used in the request to delete user attributes.}
}
\description{
Deletes the attributes for a user.
}
\examples{

}
