% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{delete_user_pool}
\alias{delete_user_pool}
\title{Deletes the specified Amazon Cognito user pool}
\usage{
delete_user_pool(
  UserPoolId = "string"
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool you want to delete.}
}
\description{
Deletes the specified Amazon Cognito user pool.
}
\examples{

}
