% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{delete_user_pool_client}
\alias{delete_user_pool_client}
\title{Allows the developer to delete the user pool client}
\usage{
delete_user_pool_client(UserPoolId, ClientId)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool where you want to delete the client.}

\item{ClientId}{The app client ID of the app associated with the user pool.}
}
\description{
Allows the developer to delete the user pool client.
}
\examples{

}
