% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{describe_user_pool}
\alias{describe_user_pool}
\title{Returns the configuration information and metadata of the specified user pool}
\usage{
describe_user_pool(UserPoolId)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool you want to describe.}
}
\description{
Returns the configuration information and metadata of the specified user pool.
}
\examples{

}
