% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{describe_user_pool_client}
\alias{describe_user_pool_client}
\title{Client method for returning the configuration information and metadata of the specified user pool app client}
\usage{
describe_user_pool_client(UserPoolId, ClientId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool you want to describe.}

\item{ClientId}{[required] The app client ID of the app associated with the user pool.}
}
\description{
Client method for returning the configuration information and metadata of the specified user pool app client.
}
\section{Accepted Parameters}{
\preformatted{describe_user_pool_client(
  UserPoolId = "string",
  ClientId = "string"
)
}
}

