% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{describe_user_pool_domain}
\alias{describe_user_pool_domain}
\title{Gets information about a domain}
\usage{
describe_user_pool_domain(Domain)
}
\arguments{
\item{Domain}{The domain string.}
}
\description{
Gets information about a domain.
}
\examples{

}
