% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{forget_device}
\alias{forget_device}
\title{Forgets the specified device}
\usage{
forget_device(DeviceKey, AccessToken = NULL)
}
\arguments{
\item{DeviceKey}{The device key.}

\item{AccessToken}{The access token for the forgotten device request.}
}
\description{
Forgets the specified device.
}
\examples{

}
