% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{get_csv_header}
\alias{get_csv_header}
\title{Gets the header information for the}
\usage{
get_csv_header(UserPoolId)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool that the users are to be imported into.}
}
\description{
Gets the header information for the .csv file to be used as input for the user import job.
}
\examples{

}
