% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{get_device}
\alias{get_device}
\title{Gets the device}
\usage{
get_device(DeviceKey, AccessToken = NULL)
}
\arguments{
\item{DeviceKey}{[required] The device key.}

\item{AccessToken}{The access token.}
}
\description{
Gets the device.
}
\section{Accepted Parameters}{
\preformatted{get_device(
  DeviceKey = "string",
  AccessToken = "string"
)
}
}

