% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{get_group}
\alias{get_group}
\title{Gets a group}
\usage{
get_group(GroupName, UserPoolId)
}
\arguments{
\item{GroupName}{[required] The name of the group.}

\item{UserPoolId}{[required] The user pool ID for the user pool.}
}
\description{
Gets a group.
}
\details{
Requires developer credentials.
}
\section{Accepted Parameters}{
\preformatted{get_group(
  GroupName = "string",
  UserPoolId = "string"
)
}
}

