% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{get_identity_provider_by_identifier}
\alias{get_identity_provider_by_identifier}
\title{Gets the specified identity provider}
\usage{
get_identity_provider_by_identifier(
  UserPoolId = "string",
  IdpIdentifier = "string"
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{IdpIdentifier}{[required] The identity provider ID.}
}
\description{
Gets the specified identity provider.
}
\examples{

}
