% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{get_ui_customization}
\alias{get_ui_customization}
\title{Gets the UI Customization information for a particular app client's app UI, if there is something set}
\usage{
get_ui_customization(
  UserPoolId = "string",
  ClientId = "string"
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{ClientId}{The client ID for the client app.}
}
\description{
Gets the UI Customization information for a particular app client's app UI, if there is something set. If nothing is set for the particular client, but there is an existing pool level customization (app \code{clientId} will be \code{ALL}), then that is returned. If nothing is present, then an empty shape is returned.
}
\examples{

}
