% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{get_user}
\alias{get_user}
\title{Gets the user attributes and metadata for a user}
\usage{
get_user(AccessToken)
}
\arguments{
\item{AccessToken}{[required] The access token returned by the server response to get information about the user.}
}
\description{
Gets the user attributes and metadata for a user.
}
\section{Accepted Parameters}{
\preformatted{get_user(
  AccessToken = "string"
)
}
}

