% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{get_user_attribute_verification_code}
\alias{get_user_attribute_verification_code}
\title{Gets the user attribute verification code for the specified attribute name}
\usage{
get_user_attribute_verification_code(AccessToken, AttributeName)
}
\arguments{
\item{AccessToken}{[required] The access token returned by the server response to get the user attribute verification code.}

\item{AttributeName}{[required] The attribute name returned by the server response to get the user attribute verification code.}
}
\description{
Gets the user attribute verification code for the specified attribute name.
}
\section{Accepted Parameters}{
\preformatted{get_user_attribute_verification_code(
  AccessToken = "string",
  AttributeName = "string"
)
}
}

