% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{global_sign_out}
\alias{global_sign_out}
\title{Signs out users from all devices}
\usage{
global_sign_out(AccessToken)
}
\arguments{
\item{AccessToken}{[required] The access token.}
}
\description{
Signs out users from all devices.
}
\section{Accepted Parameters}{
\preformatted{global_sign_out(
  AccessToken = "string"
)
}
}

