% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{list_devices}
\alias{list_devices}
\title{Lists the devices}
\usage{
list_devices(AccessToken, Limit = NULL, PaginationToken = NULL)
}
\arguments{
\item{AccessToken}{The access tokens for the request to list devices.}

\item{Limit}{The limit of the device request.}

\item{PaginationToken}{The pagination token for the list request.}
}
\description{
Lists the devices.
}
\examples{

}
