% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{list_resource_servers}
\alias{list_resource_servers}
\title{Lists the resource servers for a user pool}
\usage{
list_resource_servers(UserPoolId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{MaxResults}{The maximum number of resource servers to return.}

\item{NextToken}{A pagination token.}
}
\description{
Lists the resource servers for a user pool.
}
\section{Accepted Parameters}{
\preformatted{list_resource_servers(
  UserPoolId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

