% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{list_user_import_jobs}
\alias{list_user_import_jobs}
\title{Lists the user import jobs}
\usage{
list_user_import_jobs(UserPoolId, MaxResults, PaginationToken = NULL)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool that the users are being imported into.}

\item{MaxResults}{The maximum number of import jobs you want the request to return.}

\item{PaginationToken}{An identifier that was returned from the previous call to \code{ListUserImportJobs}, which can be used to return the next set of import jobs in the list.}
}
\description{
Lists the user import jobs.
}
\examples{

}
