% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{list_user_pools}
\alias{list_user_pools}
\title{Lists the user pools associated with an AWS account}
\usage{
list_user_pools(MaxResults, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results you want the request to return when listing the user pools.}

\item{NextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Lists the user pools associated with an AWS account.
}
\examples{

}
