% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_package.R
\docType{package}
\name{paws.cognitoidentityprovider-package}
\alias{paws.cognitoidentityprovider}
\alias{paws.cognitoidentityprovider-package}
\title{paws.cognitoidentityprovider: Amazon Cognito Identity Provider}
\description{
Using the Amazon Cognito User Pools API, you can
    create a user pool to manage directories and users. You can
    authenticate a user to obtain tokens related to user identity and
    access policies.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
