% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{resend_confirmation_code}
\alias{resend_confirmation_code}
\title{Resends the confirmation (for confirmation of registration) to a specific user in the user pool}
\usage{
resend_confirmation_code(ClientId, Username, SecretHash = NULL,
  UserContextData = NULL, AnalyticsMetadata = NULL)
}
\arguments{
\item{ClientId}{The ID of the client associated with the user pool.}

\item{Username}{The user name of the user to whom you wish to resend a confirmation code.}

\item{SecretHash}{A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.}

\item{UserContextData}{Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.}

\item{AnalyticsMetadata}{The Amazon Pinpoint analytics metadata for collecting metrics for \code{ResendConfirmationCode} calls.}
}
\description{
Resends the confirmation (for confirmation of registration) to a specific user in the user pool.
}
\examples{

}
