% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{set_user_mfa_preference}
\alias{set_user_mfa_preference}
\title{Set the user's multi-factor authentication (MFA) method preference}
\usage{
set_user_mfa_preference(AccessToken, SMSMfaSettings = NULL,
  SoftwareTokenMfaSettings = NULL)
}
\arguments{
\item{AccessToken}{The access token.}

\item{SMSMfaSettings}{The SMS text message multi-factor authentication (MFA) settings.}

\item{SoftwareTokenMfaSettings}{The time-based one-time password software token MFA settings.}
}
\description{
Set the user's multi-factor authentication (MFA) method preference.
}
\examples{

}
