% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{sign_up}
\alias{sign_up}
\title{Registers the user in the specified user pool and creates a user name, password, and user attributes}
\usage{
sign_up(ClientId, Username, Password, SecretHash = NULL,
  UserAttributes = NULL, ValidationData = NULL, AnalyticsMetadata = NULL,
  UserContextData = NULL)
}
\arguments{
\item{ClientId}{The ID of the client associated with the user pool.}

\item{Username}{The user name of the user you wish to register.}

\item{Password}{The password of the user you wish to register.}

\item{SecretHash}{A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.}

\item{UserAttributes}{An array of name-value pairs representing user attributes.

For custom attributes, you must prepend the \code{custom:} prefix to the attribute name.}

\item{ValidationData}{The validation data in the request to register a user.}

\item{AnalyticsMetadata}{The Amazon Pinpoint analytics metadata for collecting metrics for \code{SignUp} calls.}

\item{UserContextData}{Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.}
}
\description{
Registers the user in the specified user pool and creates a user name, password, and user attributes.
}
\examples{

}
