% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{start_user_import_job}
\alias{start_user_import_job}
\title{Starts the user import}
\usage{
start_user_import_job(
  UserPoolId = "string",
  JobId = "string"
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that the users are being imported into.}

\item{JobId}{[required] The job ID for the user import job.}
}
\description{
Starts the user import.
}
\examples{

}
