% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{update_device_status}
\alias{update_device_status}
\title{Updates the device status}
\usage{
update_device_status(AccessToken, DeviceKey, DeviceRememberedStatus = NULL)
}
\arguments{
\item{AccessToken}{The access token.}

\item{DeviceKey}{The device key.}

\item{DeviceRememberedStatus}{The status of whether a device is remembered.}
}
\description{
Updates the device status.
}
\examples{

}
