% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{update_group}
\alias{update_group}
\title{Updates the specified group with the specified attributes}
\usage{
update_group(GroupName, UserPoolId, Description = NULL, RoleArn = NULL,
  Precedence = NULL)
}
\arguments{
\item{GroupName}{[required] The name of the group.}

\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{Description}{A string containing the new description of the group.}

\item{RoleArn}{The new role ARN for the group. This is used for setting the \code{cognito:roles} and \code{cognito:preferred_role} claims in the token.}

\item{Precedence}{The new precedence value for the group. For more information about this parameter, see .}
}
\description{
Updates the specified group with the specified attributes.
}
\details{
Requires developer credentials.
}
\section{Accepted Parameters}{
\preformatted{update_group(
  GroupName = "string",
  UserPoolId = "string",
  Description = "string",
  RoleArn = "string",
  Precedence = 123
)
}
}

