% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{update_identity_provider}
\alias{update_identity_provider}
\title{Updates identity provider information for a user pool}
\usage{
update_identity_provider(
  UserPoolId = "string",
  ProviderName = "string",
  ProviderDetails = list(
    "string"
  ),
  AttributeMapping = list(
    "string"
  ),
  IdpIdentifiers = list(
    "string"
  )
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{ProviderName}{[required] The identity provider name.}

\item{ProviderDetails}{The identity provider details to be updated, such as \code{MetadataURL} and \code{MetadataFile}.}

\item{AttributeMapping}{The identity provider attribute mapping to be changed.}

\item{IdpIdentifiers}{A list of identity provider identifiers.}
}
\description{
Updates identity provider information for a user pool.
}
\examples{

}
