% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{verify_software_token}
\alias{verify_software_token}
\title{Use this API to register a user's entered TOTP code and mark the user's software token MFA status as "verified" if successful}
\usage{
verify_software_token(
  AccessToken = "string",
  Session = "string",
  UserCode = "string",
  FriendlyDeviceName = "string"
)
}
\arguments{
\item{AccessToken}{The access token.}

\item{Session}{The session which should be passed both ways in challenge-response calls to the service.}

\item{UserCode}{[required] The one time password computed using the secret code returned by}

\item{FriendlyDeviceName}{The friendly device name.}
}
\description{
Use this API to register a user's entered TOTP code and mark the user's software token MFA status as "verified" if successful. The request takes an access token or a session string, but not both.
}
\examples{

}
