% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{verify_user_attribute}
\alias{verify_user_attribute}
\title{Verifies the specified user attributes in the user pool}
\usage{
verify_user_attribute(AccessToken, AttributeName, Code)
}
\arguments{
\item{AccessToken}{Represents the access token of the request to verify user attributes.}

\item{AttributeName}{The attribute name in the request to verify user attributes.}

\item{Code}{The verification code in the request to verify user attributes.}
}
\description{
Verifies the specified user attributes in the user pool.
}
\examples{

}
