% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{add_custom_attributes}
\alias{add_custom_attributes}
\title{Adds additional user attributes to the user pool schema}
\usage{
add_custom_attributes(UserPoolId, CustomAttributes)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool where you want to add custom attributes.}

\item{CustomAttributes}{An array of custom attributes, such as Mutable and Name.}
}
\description{
Adds additional user attributes to the user pool schema.
}
\examples{

}
