% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_add_user_to_group}
\alias{admin_add_user_to_group}
\title{Adds the specified user to the specified group}
\usage{
admin_add_user_to_group(
  UserPoolId = "string",
  Username = "string",
  GroupName = "string"
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{Username}{[required] The username for the user.}

\item{GroupName}{[required] The group name.}
}
\description{
Adds the specified user to the specified group.
}
\details{
Requires developer credentials.
}
\examples{

}
