% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_forget_device}
\alias{admin_forget_device}
\title{Forgets the device, as an administrator}
\usage{
admin_forget_device(UserPoolId, Username, DeviceKey)
}
\arguments{
\item{UserPoolId}{The user pool ID.}

\item{Username}{The user name.}

\item{DeviceKey}{The device key.}
}
\description{
Forgets the device, as an administrator.
}
\details{
Requires developer credentials.
}
\examples{

}
