% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{admin_list_devices}
\alias{admin_list_devices}
\title{Lists devices, as an administrator}
\usage{
admin_list_devices(UserPoolId, Username, Limit = NULL,
  PaginationToken = NULL)
}
\arguments{
\item{UserPoolId}{The user pool ID.}

\item{Username}{The user name.}

\item{Limit}{The limit of the devices request.}

\item{PaginationToken}{The pagination token.}
}
\description{
Lists devices, as an administrator.
}
\details{
Requires developer credentials.
}
\examples{

}
