% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{change_password}
\alias{change_password}
\title{Changes the password for a specified user in a user pool}
\usage{
change_password(PreviousPassword, ProposedPassword, AccessToken)
}
\arguments{
\item{PreviousPassword}{[required] The old password.}

\item{ProposedPassword}{[required] The new password.}

\item{AccessToken}{[required] The access token.}
}
\description{
Changes the password for a specified user in a user pool.
}
\section{Accepted Parameters}{
\preformatted{change_password(
  PreviousPassword = "string",
  ProposedPassword = "string",
  AccessToken = "string"
)
}
}

