% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{confirm_device}
\alias{confirm_device}
\title{Confirms tracking of the device}
\usage{
confirm_device(AccessToken, DeviceKey, DeviceSecretVerifierConfig = NULL,
  DeviceName = NULL)
}
\arguments{
\item{AccessToken}{The access token.}

\item{DeviceKey}{The device key.}

\item{DeviceSecretVerifierConfig}{The configuration of the device secret verifier.}

\item{DeviceName}{The device name.}
}
\description{
Confirms tracking of the device. This API call is the call that begins device tracking.
}
\examples{

}
