% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{delete_resource_server}
\alias{delete_resource_server}
\title{Deletes a resource server}
\usage{
delete_resource_server(UserPoolId, Identifier)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that hosts the resource server.}

\item{Identifier}{[required] The identifier for the resource server.}
}
\description{
Deletes a resource server.
}
\section{Accepted Parameters}{
\preformatted{delete_resource_server(
  UserPoolId = "string",
  Identifier = "string"
)
}
}

