% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{delete_user_pool_domain}
\alias{delete_user_pool_domain}
\title{Deletes a domain for a user pool}
\usage{
delete_user_pool_domain(
  Domain = "string",
  UserPoolId = "string"
)
}
\arguments{
\item{Domain}{[required] The domain string.}

\item{UserPoolId}{[required] The user pool ID.}
}
\description{
Deletes a domain for a user pool.
}
\examples{

}
