% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{describe_identity_provider}
\alias{describe_identity_provider}
\title{Gets information about a specific identity provider}
\usage{
describe_identity_provider(UserPoolId, ProviderName)
}
\arguments{
\item{UserPoolId}{The user pool ID.}

\item{ProviderName}{The identity provider name.}
}
\description{
Gets information about a specific identity provider.
}
\examples{

}
