% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{describe_resource_server}
\alias{describe_resource_server}
\title{Describes a resource server}
\usage{
describe_resource_server(UserPoolId, Identifier)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool that hosts the resource server.}

\item{Identifier}{The identifier for the resource server}
}
\description{
Describes a resource server.
}
\examples{

}
