% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{describe_risk_configuration}
\alias{describe_risk_configuration}
\title{Describes the risk configuration}
\usage{
describe_risk_configuration(
  UserPoolId = "string",
  ClientId = "string"
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{ClientId}{The app client ID.}
}
\description{
Describes the risk configuration.
}
\examples{

}
