% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{describe_user_import_job}
\alias{describe_user_import_job}
\title{Describes the user import job}
\usage{
describe_user_import_job(UserPoolId, JobId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that the users are being imported into.}

\item{JobId}{[required] The job ID for the user import job.}
}
\description{
Describes the user import job.
}
\section{Accepted Parameters}{
\preformatted{describe_user_import_job(
  UserPoolId = "string",
  JobId = "string"
)
}
}

