% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{forget_device}
\alias{forget_device}
\title{Forgets the specified device}
\usage{
forget_device(
  AccessToken = "string",
  DeviceKey = "string"
)
}
\arguments{
\item{AccessToken}{The access token for the forgotten device request.}

\item{DeviceKey}{[required] The device key.}
}
\description{
Forgets the specified device.
}
\examples{

}
