% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{get_signing_certificate}
\alias{get_signing_certificate}
\title{This method takes a user pool ID, and returns the signing certificate}
\usage{
get_signing_certificate(UserPoolId)
}
\arguments{
\item{UserPoolId}{The user pool ID.}
}
\description{
This method takes a user pool ID, and returns the signing certificate.
}
\examples{

}
