% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{get_user_pool_mfa_config}
\alias{get_user_pool_mfa_config}
\title{Gets the user pool multi-factor authentication (MFA) configuration}
\usage{
get_user_pool_mfa_config(UserPoolId)
}
\arguments{
\item{UserPoolId}{The user pool ID.}
}
\description{
Gets the user pool multi-factor authentication (MFA) configuration.
}
\examples{

}
