% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{list_users_in_group}
\alias{list_users_in_group}
\title{Lists the users in the specified group}
\usage{
list_users_in_group(UserPoolId, GroupName, Limit = NULL, NextToken = NULL)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool.}

\item{GroupName}{The name of the group.}

\item{Limit}{The limit of the request to list users.}

\item{NextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Lists the users in the specified group.
}
\details{
Requires developer credentials.
}
\examples{

}
