% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{set_user_mfa_preference}
\alias{set_user_mfa_preference}
\title{Set the user's multi-factor authentication (MFA) method preference}
\usage{
set_user_mfa_preference(SMSMfaSettings = NULL,
  SoftwareTokenMfaSettings = NULL, AccessToken)
}
\arguments{
\item{SMSMfaSettings}{The SMS text message multi-factor authentication (MFA) settings.}

\item{SoftwareTokenMfaSettings}{The time-based one-time password software token MFA settings.}

\item{AccessToken}{[required] The access token.}
}
\description{
Set the user's multi-factor authentication (MFA) method preference.
}
\section{Accepted Parameters}{
\preformatted{set_user_mfa_preference(
  SMSMfaSettings = list(
    Enabled = TRUE|FALSE,
    PreferredMfa = TRUE|FALSE
  ),
  SoftwareTokenMfaSettings = list(
    Enabled = TRUE|FALSE,
    PreferredMfa = TRUE|FALSE
  ),
  AccessToken = "string"
)
}
}

