% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{set_user_pool_mfa_config}
\alias{set_user_pool_mfa_config}
\title{Set the user pool MFA configuration}
\usage{
set_user_pool_mfa_config(
  UserPoolId = "string",
  SmsMfaConfiguration = list(
    SmsAuthenticationMessage = "string",
    SmsConfiguration = list(
      SnsCallerArn = "string",
      ExternalId = "string"
    )
  ),
  SoftwareTokenMfaConfiguration = list(
    Enabled = TRUE|FALSE
  ),
  MfaConfiguration = "OFF"|"ON"|"OPTIONAL"
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{SmsMfaConfiguration}{The SMS text message MFA configuration.}

\item{SoftwareTokenMfaConfiguration}{The software token MFA configuration.}

\item{MfaConfiguration}{The MFA configuration.}
}
\description{
Set the user pool MFA configuration.
}
\examples{

}
