% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{stop_user_import_job}
\alias{stop_user_import_job}
\title{Stops the user import job}
\usage{
stop_user_import_job(UserPoolId, JobId)
}
\arguments{
\item{UserPoolId}{The user pool ID for the user pool that the users are being imported into.}

\item{JobId}{The job ID for the user import job.}
}
\description{
Stops the user import job.
}
\examples{

}
