% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentityprovider_operations.R
\name{update_user_pool}
\alias{update_user_pool}
\title{Updates the specified user pool with the specified attributes}
\usage{
update_user_pool(UserPoolId, Policies = NULL, LambdaConfig = NULL,
  AutoVerifiedAttributes = NULL, SmsVerificationMessage = NULL,
  EmailVerificationMessage = NULL, EmailVerificationSubject = NULL,
  VerificationMessageTemplate = NULL, SmsAuthenticationMessage = NULL,
  MfaConfiguration = NULL, DeviceConfiguration = NULL,
  EmailConfiguration = NULL, SmsConfiguration = NULL,
  UserPoolTags = NULL, AdminCreateUserConfig = NULL,
  UserPoolAddOns = NULL)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool you want to update.}

\item{Policies}{A container with the policies you wish to update in a user pool.}

\item{LambdaConfig}{The AWS Lambda configuration information from the request to update the user pool.}

\item{AutoVerifiedAttributes}{The attributes that are automatically verified when the Amazon Cognito service makes a request to update user pools.}

\item{SmsVerificationMessage}{A container with information about the SMS verification message.}

\item{EmailVerificationMessage}{The contents of the email verification message.}

\item{EmailVerificationSubject}{The subject of the email verification message.}

\item{VerificationMessageTemplate}{The template for verification messages.}

\item{SmsAuthenticationMessage}{The contents of the SMS authentication message.}

\item{MfaConfiguration}{Can be one of the following values:
\itemize{
\item \code{OFF} - MFA tokens are not required and cannot be specified during user registration.
\item \code{ON} - MFA tokens are required for all user registrations. You can only specify required when you are initially creating a user pool.
\item \code{OPTIONAL} - Users have the option when registering to create an MFA token.
}}

\item{DeviceConfiguration}{Device configuration.}

\item{EmailConfiguration}{Email configuration.}

\item{SmsConfiguration}{SMS configuration.}

\item{UserPoolTags}{The cost allocation tags for the user pool. For more information, see \href{http://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-pools-cost-allocation-tagging.html}{Adding Cost Allocation Tags to Your User Pool}}

\item{AdminCreateUserConfig}{The configuration for \code{AdminCreateUser} requests.}

\item{UserPoolAddOns}{Used to enable advanced security risk detection. Set the key \code{AdvancedSecurityMode} to the value "AUDIT".}
}
\description{
Updates the specified user pool with the specified attributes. If you don't provide a value for an attribute, it will be set to the default value. You can get a list of the current user pool settings with .
}
\section{Accepted Parameters}{
\preformatted{update_user_pool(
  UserPoolId = "string",
  Policies = list(
    PasswordPolicy = list(
      MinimumLength = 123,
      RequireUppercase = TRUE|FALSE,
      RequireLowercase = TRUE|FALSE,
      RequireNumbers = TRUE|FALSE,
      RequireSymbols = TRUE|FALSE
    )
  ),
  LambdaConfig = list(
    PreSignUp = "string",
    CustomMessage = "string",
    PostConfirmation = "string",
    PreAuthentication = "string",
    PostAuthentication = "string",
    DefineAuthChallenge = "string",
    CreateAuthChallenge = "string",
    VerifyAuthChallengeResponse = "string",
    PreTokenGeneration = "string",
    UserMigration = "string"
  ),
  AutoVerifiedAttributes = list(
    "phone_number"|"email"
  ),
  SmsVerificationMessage = "string",
  EmailVerificationMessage = "string",
  EmailVerificationSubject = "string",
  VerificationMessageTemplate = list(
    SmsMessage = "string",
    EmailMessage = "string",
    EmailSubject = "string",
    EmailMessageByLink = "string",
    EmailSubjectByLink = "string",
    DefaultEmailOption = "CONFIRM_WITH_LINK"|"CONFIRM_WITH_CODE"
  ),
  SmsAuthenticationMessage = "string",
  MfaConfiguration = "OFF"|"ON"|"OPTIONAL",
  DeviceConfiguration = list(
    ChallengeRequiredOnNewDevice = TRUE|FALSE,
    DeviceOnlyRememberedOnUserPrompt = TRUE|FALSE
  ),
  EmailConfiguration = list(
    SourceArn = "string",
    ReplyToEmailAddress = "string"
  ),
  SmsConfiguration = list(
    SnsCallerArn = "string",
    ExternalId = "string"
  ),
  UserPoolTags = list(
    "string"
  ),
  AdminCreateUserConfig = list(
    AllowAdminCreateUserOnly = TRUE|FALSE,
    UnusedAccountValidityDays = 123,
    InviteMessageTemplate = list(
      SMSMessage = "string",
      EmailMessage = "string",
      EmailSubject = "string"
    )
  ),
  UserPoolAddOns = list(
    AdvancedSecurityMode = "OFF"|"AUDIT"|"ENFORCED"
  )
)
}
}

